Move ObjBase and CollectionBase to legacy. Added deprecation message (#245)
