+++ src/easyscience/__init__.py
@@ -14,6 +14,7 @@ global_object.stack.enabled = False
 from .base_classes import ObjBase  # noqa: E402
 from .fitting import AvailableMinimizers  # noqa: E402
 from .fitting import Fitter  # noqa: E402
+from .legacy import CollectionBase  # noqa: E402
 from .variable import DescriptorNumber  # noqa: E402
 from .variable import Parameter  # noqa: E402
 
@@ -23,6 +24,7 @@ __all__ = [
     __version__,
     global_object,
     ObjBase,
+    CollectionBase,
     AvailableMinimizers,
     Fitter,
     DescriptorNumber,
