+++ src/easyscience/base_classes/__init__.py
@@ -1,11 +1,11 @@
 # SPDX-FileCopyrightText: 2025 EasyScience contributors <https://github.com/easyscience>
 # SPDX-License-Identifier: BSD-3-Clause
 
+from ..legacy.collection_base import CollectionBase
+from ..legacy.obj_base import ObjBase
 from .based_base import BasedBase
-from .collection_base import CollectionBase
 from .easy_list import EasyList
 from .model_base import ModelBase
 from .new_base import NewBase
-from .obj_base import ObjBase
 
 __all__ = [BasedBase, CollectionBase, ObjBase, ModelBase, NewBase, EasyList]
