+++ src/easyscience/job/analysis.py
@@ -6,8 +6,8 @@ from abc import abstractmethod
 
 import numpy as np
 
-from ..base_classes.obj_base import ObjBase
 from ..fitting.minimizers import MinimizerBase
+from ..legacy.obj_base import ObjBase
 
 
 class AnalysisBase(ObjBase, metaclass=ABCMeta):
