+++ src/easyscience/job/experiment.py
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: 2026 EasyScience contributors <https://github.com/easyscience>
 # SPDX-License-Identifier: BSD-3-Clause
 
-from ..base_classes.obj_base import ObjBase
+from ..legacy.obj_base import ObjBase
 
 
 class ExperimentBase(ObjBase):
